requirements.txt
streamlit==1.28.1
pillow==10.0.0
pandas==2.0.3
sqlite3
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0